fix: cast vote sum to int in backfill update to match compute_hot_score signature

+++ supabase/migrations/20251027005237_cleanup_perf.sql
@@ -86,7 +86,7 @@ execute function public.votes_hot_trg_router();
 -- 4) BACKFILL HOT SCORE (vetëm kur është null) ------------
 update public.posts p
 set hot_score = public.compute_hot_score(
-  coalesce( (select sum(value) from public.votes v where v.post_id = p.id), 0 ),
+  coalesce( (select sum(value) from public.votes v where v.post_id = p.id), 0 )::int,
   p.created_at
 )
 where p.hot_score is null;
